+++ app/hub/Views/SetupView/__tests__/SetupViewContainer.test.jsx
@@ -45,7 +45,6 @@ const actions = {
 	setAntiTracking: () => {},
 	setAdBlock: () => {},
 	setSmartBlocking: () => {},
-	setGhosteryRewards: () => {},
 	setHumanWeb: () => {},
 	setSetupComplete: () => {},
 };