+++ app/shared-components/PromoModal/index.js
@@ -11,6 +11,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0
  */
 
+import { connect } from 'react-redux';
 import PromoModalContainer from './PromoModalContainer';
 
-export default PromoModalContainer;
+// We may be in the Hub, where state.panel may be undefined
+const mapStateToProps = state => ({ tab_id: ((state.panel && state.panel.tab_id) || null) });
+
+export default connect(mapStateToProps, undefined)(PromoModalContainer);